Refactored: MongoDB instrumentation into module.
